working version without test object in animate, bullets not visible in 2nd window
